BattleCharacterEntityを新規構築するユースケースを作成する (#131)
* AdjustPlayerBattleEntityUsecaseを作成

* AdjustPlayerBattleEntityUsecaseのテストを追加

* update package json

+++ Assets/Plugins/BlackSmith.Domain/Usecase/Character/AdjustPlayerCommonEntityUsecase.cs
@@ -4,16 +4,18 @@ using BlackSmith.Usecase.Interface;
 using Cysharp.Threading.Tasks;
 using System;
 
+#nullable enable
+
 namespace BlackSmith.Usecase.Character
 {
     /// <summary>
     /// CommonPlayerEntityの作成、削除を行うユースケース
     /// </summary>
-    public class AdjustPlayerCommonUsecase
+    public class AdjustPlayerCommonEntityUsecase
     {
         private readonly IPlayerCommonEntityRepository repository;
 
-        public AdjustPlayerCommonUsecase(IPlayerCommonEntityRepository playerRepository)
+        public AdjustPlayerCommonEntityUsecase(IPlayerCommonEntityRepository playerRepository)
         {
             repository = playerRepository;
         }
@@ -55,23 +57,24 @@ namespace BlackSmith.Usecase.Character
         /// </summary>
         /// <param name="id"></param>
         /// <returns></returns>
-        /// <exception cref="ArgumentException">指定したIdのキャラクターが存在しない場合</exception>
-        public async UniTask<PlayerCommonEntity> GetCharacter(CharacterID id)
+        public async UniTask<PlayerCommonEntity?> GetCharacter(CharacterID id)
         {
             if (!await repository.IsExist(id))
-                throw new ArgumentException($"指定したidのキャラクターは存在しません {id}");
+                return null;
 
-            var entity = (await repository.FindByID(id)) ?? throw new ArgumentException($"指定したidのキャラクターは存在しません {id}");
-
-            return entity;
+            return await repository.FindByID(id);
         }
 
         /// <summary>
         /// プレイヤーデータの削除を行う
         /// </summary>
         /// <param name="id">削除するプレイヤーのID</param>
+        /// <exception cref="InvalidOperationException">指定したidのキャラクターは存在しない場合</exception>
         public async UniTask DeletePlayer(CharacterID id)
         {
+            if (!await repository.IsExist(id))
+                throw new InvalidOperationException($"指定したidのキャラクターは存在しません {id}");
+
             await repository.Delete(id);
         }
     }
